+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Upcoming
+* Experimental Flying blocks explosion changes 
+  * Removed smoke particles
+  * Reduced knockback taken by flying blocks
+  * Fixed Flying blocks not flying
+
 ## 2.12.4
 * Tiredness
   * Added Energy Boost